copilot/backend/requirements.txt
# Trade Compliance Backend Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
snowflake-connector-python>=3.6.0
python-multipart>=0.0.6
sse-starlette>=1.8.2
httpx>=0.26.0

# ML Dependencies (for notebooks)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.18.0

# Development
python-dotenv>=1.0.0
